+++ src/display/stubs.js
@@ -13,6 +13,8 @@
  * limitations under the License.
  */
 
+const DOMCMapReaderFactory = null;
+const DOMStandardFontDataFactory = null;
 const NodeCanvasFactory = null;
 const NodeCMapReaderFactory = null;
 const NodeFilterFactory = null;
@@ -23,6 +25,8 @@ const PDFNetworkStream = null;
 const PDFNodeStream = null;
 
 export {
+  DOMCMapReaderFactory,
+  DOMStandardFontDataFactory,
   NodeCanvasFactory,
   NodeCMapReaderFactory,
   NodeFilterFactory,
