+++ test/unit/clitests.json
@@ -9,6 +9,7 @@
     "api_spec.js",
     "app_options_spec.js",
     "bidi_spec.js",
+    "canvas_factory_spec.js",
     "cff_parser_spec.js",
     "cmap_spec.js",
     "colorspace_spec.js",
@@ -42,6 +43,7 @@
     "primitives_spec.js",
     "stream_spec.js",
     "struct_tree_spec.js",
+    "svg_factory_spec.js",
     "text_layer_spec.js",
     "type1_parser_spec.js",
     "ui_utils_spec.js",
