+++ test/unit/unit_test.html
@@ -20,6 +20,8 @@
         "fluent-dom": "../../node_modules/@fluent/dom/esm/index.js",
         "cached-iterable": "../../node_modules/cached-iterable/src/index.mjs",
 
+        "display-cmap_reader_factory": "../../src/display/cmap_reader_factory.js",
+        "display-standard_fontdata_factory": "../../src/display/standard_fontdata_factory.js",
         "display-fetch_stream": "../../src/display/fetch_stream.js",
         "display-network": "../../src/display/network.js",
         "display-node_stream": "../../src/display/stubs.js",
