+++ web/viewer-geckoview.html
@@ -59,6 +59,8 @@ See https://github.com/adobe-type-tools/cmap-resources
           "fluent-dom": "../node_modules/@fluent/dom/esm/index.js",
           "cached-iterable": "../node_modules/cached-iterable/src/index.mjs",
 
+          "display-cmap_reader_factory": "../src/display/cmap_reader_factory.js",
+          "display-standard_fontdata_factory": "../src/display/standard_fontdata_factory.js",
           "display-fetch_stream": "../src/display/fetch_stream.js",
           "display-network": "../src/display/network.js",
           "display-node_stream": "../src/display/stubs.js",
